Add config file for configuring the creation of the amalgamation.

+++ create_amalgamation.py
@@ -1,36 +1,38 @@
 import collections
 import itertools
+import json
+import os.path
 import re
+import sys
+
+# Load the config file (required)
+if not os.path.isfile('ckit.config'):
+    print 'ckit.config file is missing, please create it.'
+    sys.exit(1)
+with open('ckit.config', 'r') as f:
+    config = json.load(f)
 
 # The list of standard include files that we will need
 std_incs = 'assert math stdio stdlib string'.split()
 
-# Other includes
-oth_incs = 'test_struct'.split()
-
 # The print regular expression
 print_re = re.compile(r'\$print\{(.*)}')
 
-# The instances of data structures we need to generate
-Inst = collections.namedtuple('Inst', 'type_ print_')
-instances = {'array': [Inst('int', r'printf("%d", \1)'), Inst('test_struct', r'print_test_struct(\1)')],
-             'heap' : [Inst('int', r'printf("%d", \1)')]}
-
 # Read all input files
 lines = dict()
-for name in instances:
-    lines[name] = dict()
+for compname in config['components']:
+    lines[compname] = dict()
     for ext in 'c h'.split():
-        with open(name + '.' + ext, 'r') as f:
-            lines[name][ext] = f.readlines()
+        with open(compname + '.' + ext, 'r') as f:
+            lines[compname][ext] = f.readlines()
 
 # Write output files
 with open('ckit.h', 'w') as outf_h, open('ckit.c', 'w') as outf_c:
 
     # Write initial stuff to both files
     outf_h.write('#ifndef CKIT_H\n#define CKIT_H\n\n')
-    for inc in oth_incs:
-        outf_h.write('#include "{}.h"\n'.format(inc))
+    for inc in config['includes']:
+        outf_h.write('#include "{}"\n'.format(inc))
     outf_c.write('#include "ckit.h"\n\n')
     for inc in std_incs:
         outf_c.write('#include <{}.h>\n'.format(inc))
@@ -38,12 +40,12 @@ with open('ckit.h', 'w') as outf_h, open('ckit.c', 'w') as outf_c:
 
     # Write actual content
     outf = {'c': outf_c, 'h': outf_h}
-    for name, insts in instances.iteritems():
-        for instance in insts:
+    for compname, typenames in config['components'].iteritems():
+        for typename in typenames:
             for ext, f in outf.iteritems():
-                for line in lines[name][ext]:
-                    line = line.replace('$type', instance.type_)
-                    line = print_re.sub(instance.print_, line)
+                for line in lines[compname][ext]:
+                    line = line.replace('$type', typename)
+                    line = print_re.sub(config['types'][typename]['print'], line)
                     f.write(line)
 
     # Write final stuff to both files
